Module Yocaml_runtime

module Error : sig ... end

Description of common errors in the various runtimes (at least shared by Unix and Eio).

module Hash : sig ... end

Helpers for hashing, in SHA256, strings

module Log : sig ... end

Utilities for working with the Logs library in a given runtime.

module Server : sig ... end

Generic tools to facilitate server construction (component sharing, etc.)