Module Yocaml_runtime.Error

Description of common errors in the various runtimes (at least shared by Unix and Eio).

Types

type common =

| Unable_to_write_file of Yocaml.Path.t * string

| Unable_to_create_directory of Yocaml.Path.t

| Unable_to_read_file of Yocaml.Path.t

| Unable_to_read_directory of Yocaml.Path.t

| Unable_to_read_mtime of Yocaml.Path.t

| Unable_to_perform_command of string * exn

Describes common errors that can occur during the execution of a program in a specific runtime.

Utils

val common_to_string : common -> string

String representation of a common.