Module Yocaml_runtime.Log

Utilities for working with the Logs library in a given runtime.

Types

type level = [

| `App

| `Error

| `Warning

| `Info

| `Debug

]

Describe a log-level.

Helpers

val msg : ?src:Logs.src -> level -> string -> unit

msg level message log a message with a given message.

val setup : ?level:level -> unit -> unit

Set-up a default logger.